Update RNK and BSN check user Update DB.

+++ trunk/Sources/Source_Codes/FBDSource/FBD/Controllers/BSNIndustryController.cs
@@ -23,6 +23,11 @@ namespace FBD.Controllers
         
         public ActionResult Index()
         {
+            if (!AccessManager.AllowAccess(Constants.RIGHT_PARAMETERS_VIEW, Session[Constants.SESSION_USER_ID]))
+            {
+                return RedirectToAction("Unauthorized", "SYSAuths");
+            }
+
             List<BusinessIndustries> industries=null;
             try
             {
@@ -49,6 +54,10 @@ namespace FBD.Controllers
         /// <returns></returns>
         public ActionResult Add()
         {
+            if (!AccessManager.AllowAccess(Constants.RIGHT_PARAMETERS_UPDATE, Session[Constants.SESSION_USER_ID]))
+            {
+                return RedirectToAction("Unauthorized", "SYSAuths");
+            }
             return View();
         } 
 
@@ -62,6 +71,10 @@ namespace FBD.Controllers
         [HttpPost]
         public ActionResult Add(BusinessIndustries industry)
         {
+            if (!AccessManager.AllowAccess(Constants.RIGHT_PARAMETERS_UPDATE, Session[Constants.SESSION_USER_ID]))
+            {
+                return RedirectToAction("Unauthorized", "SYSAuths");
+            }
             try
             {
                 if (ModelState.IsValid)
@@ -98,6 +111,10 @@ namespace FBD.Controllers
          /// <returns></returns>
         public ActionResult Edit(string id)
         {
+            if (!AccessManager.AllowAccess(Constants.RIGHT_PARAMETERS_UPDATE, Session[Constants.SESSION_USER_ID]))
+            {
+                return RedirectToAction("Unauthorized", "SYSAuths");
+            }
             BusinessIndustries model = null;
             try
             {
@@ -125,6 +142,10 @@ namespace FBD.Controllers
         [HttpPost]
         public ActionResult Edit(string id, BusinessIndustries industry)
         {
+            if (!AccessManager.AllowAccess(Constants.RIGHT_PARAMETERS_UPDATE, Session[Constants.SESSION_USER_ID]))
+            {
+                return RedirectToAction("Unauthorized", "SYSAuths");
+            }
             try
             {
                 
@@ -159,6 +180,10 @@ namespace FBD.Controllers
          /// <returns></returns>
         public ActionResult Delete(string id)
         {
+            if (!AccessManager.AllowAccess(Constants.RIGHT_PARAMETERS_UPDATE, Session[Constants.SESSION_USER_ID]))
+            {
+                return RedirectToAction("Unauthorized", "SYSAuths");
+            }
             try
             {
                 int result=BusinessIndustries.DeleteIndustry(id);
